Add many services

+++ src/main/kotlin/com/radioteria/domain/service/ChannelTracklistService.kt
@@ -5,5 +5,5 @@ import com.radioteria.domain.entity.Track
 interface ChannelTracklistService {
     fun add(track: Track)
     fun delete(track: Track)
-    fun move(track: Track, newPosition: Long)
+    fun move(track: Track, newPosition: Int)
 }
